requirements.txt
confusable-homoglyphs==3.2.0
Django>=3.2.13,==3.2.*
django-appconf==1.0.5
django-axes==5.35
django-ipware==4.0.2
django-registration==3.3
django-simple-captcha==0.5.17
gitdb2>=4.0.2,==4.0.*
GitPython==3.1.27
gunicorn==20.1.0
mysqlclient==2.1.0
pytz==2022.1
six==1.16.0
smmap2==3.0.1
Pillow==9.1.1 # not directly required, pinned by Snyk to avoid a vulnerability
django-csp==3.7
